+++ 05-examples/invoice/02-bpmn/src/main/scala/camundala/examples/invoice/bpmn/ReviewInvoice.scala
@@ -6,6 +6,8 @@ import camundala.domain.*
 object ReviewInvoice extends BpmnDsl:
   final val processName = "example-invoice-c7-review"
 
+  type InConfig = NoInConfig
+
   @description("Same Input as _InvoiceReceipt_, only different Mocking")
   case class In(
       creditor: String = "Great Pizza for Everyone Inc.",
@@ -27,14 +29,12 @@ object ReviewInvoice extends BpmnDsl:
     given InOutCodec[Out] = deriveCodec
   end Out
 
-  lazy val example: Process[In, Out] =
+  lazy val example =
     process(
       id = processName,
-      descr = // cawemoDescr(
-        "This starts the Review Invoice Process.",
-      // "cc9f978a-e98a-4b01-991d-36d682574cda"),
+      descr = "This starts the Review Invoice Process.",
       in = In(),
-      out = Out()
+      out = Out(),
     )
 
   object AssignReviewerUT: