+++ 05-examples/invoice/03-simulation/src/test/scala/camundala/examples/invoice/simulation/InvoiceSimulation.scala
@@ -13,11 +13,11 @@ import camundala.simulation.custom.CustomSimulation
 class InvoiceSimulation extends CustomSimulation:
 
   simulate(
-    ignore.scenario(`Review Invoice`)(
+    scenario(`Review Invoice`)(
       AssignReviewerUT,
       ReviewInvoiceUT
     ),
-    ignore.incidentScenario(
+    incidentScenario(
       `Invoice Receipt that fails`,
       "Could not archive invoice..."
     )(
@@ -30,11 +30,11 @@ class InvoiceSimulation extends CustomSimulation:
       ApproveInvoiceUT,
       PrepareBankTransferUT
     ),
-    ignore.scenario(WithOverrideScenario)(
+    scenario(WithOverrideScenario)(
       `ApproveInvoiceUT with Override`,
       PrepareBankTransferUT
     ),
-    ignore.scenario(`Invoice Receipt with Review`)(
+    scenario(`Invoice Receipt with Review`)(
       NotApproveInvoiceUT,
       subProcess(`Review Invoice`)(
         AssignReviewerUT,
@@ -43,65 +43,67 @@ class InvoiceSimulation extends CustomSimulation:
       ApproveInvoiceUT, // now approve
       PrepareBankTransferUT
     ),
-    ignore.scenario(`Invoice Receipt with Review failed`)(
+    scenario(`Invoice Receipt with Review failed`)(
       NotApproveInvoiceUT, // do not approve
       subProcess(`Review Invoice not clarified`)(
         AssignReviewerUT,
         ReviewInvoiceNotClarifiedUT // do not clarify
       )
     ),
-    ignore.scenario(`Invoice Receipt with Review mocked`)(
+    scenario(`Invoice Receipt with Review mocked`)(
       NotApproveInvoiceUT,
       // subProcess Mocked - so nothing to do
       ApproveInvoiceUT, // now approve
       PrepareBankTransferUT
     ),
-    ignore.scenario(InvoiceAssignApproverDMN),
-    ignore.scenario(InvoiceAssignApproverDMN2),
-    ignore.incidentScenario(
+    scenario(InvoiceAssignApproverDMN),
+    scenario(InvoiceAssignApproverDMN2),
+    incidentScenario(
       BadValidationP,
       "DecodingFailure at .creditor: Got value 'null' with wrong type, expecting string"
     ),
     // mocking
-    ignore.scenario(`Invoice Receipt mocked invoiceReviewed`)(
+    scenario(`Invoice Receipt mocked invoiceReviewed`)(
       NotApproveInvoiceUT,
       // subProcess not needed because of mocking
       ApproveInvoiceUT, // now approve
       PrepareBankTransferUT
     ),
-    ignore.scenario(`Review Invoice mocked`), // mocks itself
+    scenario(`Review Invoice mocked`), // mocks itself
     // ExternalTask - works only if exampleInvoiceWorkerC7 is running
-    ignore.scenario(
+    scenario(
       `Archive Invoice`
     ),
-    ignore.scenario(
+    scenario(
       `Archive Invoice handled`
     ),
-    ignore.scenario(
+    scenario(
       `Archive Invoice handled regex matched`
     ),
-    ignore.incidentScenario(
+    incidentScenario(
       `Archive Invoice handled not matched`,
       "The error was handled, but did not match the defined 'regexHandledErrors'."
     ),
-    ignore.incidentScenario(
+    incidentScenario(
       `Archive Invoice that fails`,
       "Could not archive invoice"
     ),
-    ignore.scenario(`Archive Invoice mocked output`),
+    scenario(`Archive Invoice mocked output`)
   )
 
   override implicit def config =
     super.config
       .withPort(8034)
-      //.withLogLevel(LogLevel.DEBUG)
+    // .withMaxCount(20)
+    // .withLogLevel(LogLevel.DEBUG)
 
   private lazy val `Invoice Receipt` = InvoiceReceipt.example
   private lazy val `Invoice Receipt mocked invoiceReviewed` =
     `Invoice Receipt with Review`
-      .withIn(
-        InvoiceReceipt.In(invoiceReviewedMock = Some(ReviewInvoice.Out()))
-      )
+      .withIn(InvoiceReceipt.In(inConfig =
+        Some(InvoiceReceipt.InConfig(
+          invoiceReviewedMock = Some(ReviewInvoice.Out())))
+      ))
 
   private lazy val ApproveInvoiceUT = InvoiceReceipt.ApproveInvoiceUT.example
   private lazy val PrepareBankTransferUT =
@@ -138,7 +140,7 @@ class InvoiceSimulation extends CustomSimulation:
       .isEquals("amount", 300.0)
   private lazy val `Invoice Receipt that fails` =
     InvoiceReceipt.example
-      .withIn(InvoiceReceipt.In(shouldFail = Some(true)))
+      .withIn(InvoiceReceipt.In(inConfig = Some(InvoiceReceipt.InConfig(shouldFail = Some(true)))))
 
   private lazy val InvoiceAssignApproverDMN =
     InvoiceReceipt.InvoiceAssignApproverDMN.example
@@ -200,6 +202,5 @@ class InvoiceSimulation extends CustomSimulation:
   private lazy val `Archive Invoice mocked output` =
     ArchiveInvoice.example
       .mockWith(ArchiveInvoice.Out(Some(true)))
-  
-    
+
 end InvoiceSimulation